pyproject.toml
[tool.poetry]
name = "template"
version = "0.1.0"
description = "Multi-Agent System Planner using AutoGen"
authors = ["BaoBao <baobao@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
psycopg2-binary = "^2.9.10"
fastapi = "^0.112.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.7.1"
apscheduler = "^3.11.0"
pyautogen = "^0.4.0"
google-cloud-aiplatform = "^1.38.1"
APScheduler = "^3.11.0"
python-telegram-bot = "^21.10"
googlesearch-python = "^1.3.0"
dexscreener = "^1.2"
mem0ai = "^0.1.93"
timeout-decorator = "^0.5.0"
fake_useragent = "^1.5.1"
python-dotenv = "^1.0.0"
redis = "^6.4.0"
termcolor = "^2.3.0"
langchain-mcp-adapters="0.0.9"
python-multipart="0.0.20"
openai = "^1.0.0"
aiofiles = "^23.0.0"
httpx = "^0.27.0"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
